+++ R/lookup_by_column.R
@@ -78,7 +78,8 @@ lookup_by_column <- function(source,
   f_db <- function(db,
                    incoming_table_name,
                    outgoing_table_name,
-                   nd = NULL) {
+                   nd = NULL,
+                   ...) {
     # get list of possible values
     q <- paste0("
      SELECT