+++ R/quantile.R
@@ -165,7 +165,8 @@ quantile_node <- function(source,
   f_db <- function(db,
                    incoming_table_name,
                    outgoing_table_name,
-                   nd = NULL) {
+                   nd = NULL,
+                   ...) {
     qtable <- quantile_cols(db, incoming_table_name,
                             probs = probs,
                             probs_name = probs_name,